refactoring main algorithm code
